requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
